+++ base/test_helpers.go
@@ -1,0 +1,15 @@
+package base
+
+import (
+	"testing"
+
+	"github.com/jarcoal/httpmock"
+)
+
+func SetupWithMock(t *testing.T) {
+	httpmock.Activate()
+	// Exact URL match
+	httpmock.RegisterResponder("GET", "https://httpbin.org/get",
+		httpmock.NewStringResponder(200, `[{"id": 1, "name": "My Great Thing"}]`))
+	t.Cleanup(httpmock.DeactivateAndReset)
+}
