+++ testdata/drivertests/experiment.yaml
@@ -1,0 +1,39 @@
+# task 1: generate HTTP requests for application URL
+# collect Iter8's built-in HTTP latency and error-related metrics
+- task: gen-load-and-collect-metrics-http
+  with:
+    duration: 2s
+    errorRanges:
+    - lower: 500
+    percentiles: 
+    - 50
+    - 75
+    - 90
+    - 95
+    - 97.5
+    - 99
+    - 99.9
+    versionInfo:
+    - url: https://httpbin.org/get
+# task 2: validate service level objectives for app using
+# the metrics collected in the above task
+- task: assess-app-versions
+  with:
+    SLOs:
+    - metric: "http/error-rate"
+      upperLimit: 0
+    - metric: "http/latency-mean"
+      upperLimit: 500
+    - metric: "http/latency-p50"
+      upperLimit: 1000
+    - metric: "http/latency-p50.0"
+      upperLimit: 1000
+    - metric: "http/latency-p95.0"
+      upperLimit: 2500
+    - metric: "http/latency-p97.5"
+      upperLimit: 5000
+# tasks 3 & 4: print if SLOs are satisfied or not
+- if: SLOs()
+  run: echo "SLOs satisfied"
+- if: not SLOs()
+  run: echo "SLOs not satisfied"
